update fabric for python3 support
